header for resizearray
